+++ mars/services/task/execution/ray/tests/test_ray_execution_backend.py
@@ -12,16 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections import Counter
-
-import numpy as np
+import asyncio
 import pandas as pd
 import pytest
+import numpy as np
 
+from collections import Counter
+from ....analyzer import GraphAnalyzer
+from .....subtask import SubtaskGraph
 from ...... import tensor as mt
+from ......config import Config
 
-from ......core import TileContext
+from ......core import TileContext, ChunkGraph
 from ......core.graph import TileableGraph, TileableGraphBuilder, ChunkGraphBuilder
+
+from ......core.operand import Fetch
+
+from ......resource import Resource
 from ......serialization import serialize
 from ......tests.core import require_ray, mock
 from ......utils import lazy_import, get_chunk_params
@@ -46,15 +53,21 @@ def _gen_subtask_chunk_graph(t):
     return next(ChunkGraphBuilder(graph, fuse_enabled=False).build())
 
 
+def _gen_subtask_graph(t):
+    tileable_graph = t.build_graph(tile=False)
+    chunk_graph = next(ChunkGraphBuilder(tileable_graph).build())
+    bands = [(f"address_{i}", "numa-0") for i in range(4)]
+    band_resource = dict((band, Resource(num_cpus=1)) for band in bands)
+    task = Task("mock_task", "mock_session", tileable_graph)
+    analyzer = GraphAnalyzer(chunk_graph, band_resource, task, Config(), dict())
+    subtask_graph = analyzer.gen_subtask_graph()
+    return chunk_graph, subtask_graph
+
+
 class MockRayTaskExecutor(RayTaskExecutor):
     def __init__(self, *args, **kwargs):
         self._set_attrs = Counter()
         super().__init__(*args, **kwargs)
-
-    @staticmethod
-    def _get_ray_executor():
-        # Export remote function once.
-        return None
 
     def set_attr_counter(self):
         return self._set_attrs
@@ -62,6 +75,61 @@ class MockRayTaskExecutor(RayTaskExecutor):
     def __setattr__(self, key, value):
         super().__setattr__(key, value)
         self._set_attrs[key] += 1
+
+    async def submit_subtask_graph(
+        self,
+        stage_id: str,
+        subtask_graph: SubtaskGraph,
+        chunk_graph: ChunkGraph,
+    ):
+        monitor_task = asyncio.create_task(
+            self._update_progress_and_collect_garbage(
+                subtask_graph, self._config.get_subtask_monitor_interval()
+            )
+        )
+
+        result_meta_keys = {
+            chunk.key
+            for chunk in chunk_graph.result_chunks
+            if not isinstance(chunk.op, Fetch)
+        }
+
+        for subtask in subtask_graph.topological_iter():
+            subtask_chunk_graph = subtask.chunk_graph
+            task_context = self._task_context
+            key_to_input = await self._load_subtask_inputs(
+                stage_id, subtask, subtask_chunk_graph, task_context
+            )
+            output_keys = self._get_subtask_output_keys(subtask_chunk_graph)
+            output_meta_keys = result_meta_keys & output_keys
+            output_count = len(output_keys) + bool(output_meta_keys)
+            output_object_refs = self._ray_executor.options(
+                num_returns=output_count
+            ).remote(
+                subtask.task_id,
+                subtask.subtask_id,
+                serialize(subtask_chunk_graph),
+                output_meta_keys,
+                list(key_to_input.keys()),
+                *key_to_input.values(),
+            )
+            if output_count == 0:
+                continue
+            elif output_count == 1:
+                output_object_refs = [output_object_refs]
+            self._cur_stage_first_output_object_ref_to_subtask[
+                output_object_refs[0]
+            ] = subtask
+            if output_meta_keys:
+                meta_object_ref, *output_object_refs = output_object_refs
+            task_context.update(zip(output_keys, output_object_refs))
+
+        return monitor_task
+
+
+class MockTileContext(TileContext):
+    def get_all_progress(self) -> float:
+        return 1.0
 
 
 @pytest.mark.asyncio
@@ -219,3 +287,62 @@ def test_ray_execution_worker_context():
     assert context.set_progress(0.1) is None
     assert context.new_custom_log_dir() is None
     assert context.get_storage_info("mock_address") == {}
+
+
+@require_ray
+@pytest.mark.asyncio
+async def test_executor_context_gc():
+    popped_seq = []
+
+    class MockTaskContext(dict):
+        def pop(self, k, d=None):
+            popped_seq.append(k)
+            return super().pop(k, d)
+
+    t1 = mt.random.randint(10, size=(100, 10), chunk_size=100)
+    t2 = mt.random.randint(10, size=(100, 10), chunk_size=50)
+    t3 = t2 + t1
+    t4 = t3.sum(0)
+    chunk_graph, subtask_graph = _gen_subtask_graph(t4)
+    task = Task("mock_task", "mock_session", fuse_enabled=True)
+    mock_config = RayExecutionConfig.from_execution_config(
+        {"backend": "ray", "ray": {"subtask_monitor_interval": 0}}
+    )
+    tile_context = MockTileContext()
+    task_context = MockTaskContext()
+    executor = MockRayTaskExecutor(
+        config=mock_config,
+        task=task,
+        tile_context=tile_context,
+        task_context=task_context,
+        task_chunks_meta={},
+        task_state_actor=None,
+        lifecycle_api=None,
+        meta_api=None,
+    )
+    monitor_task = await executor.submit_subtask_graph(
+        "mock_stage", subtask_graph, chunk_graph
+    )
+    await monitor_task
+
+    assert len(task_context) == 1
+    assert len(popped_seq) == 6
+    subtasks = list(subtask_graph.topological_iter())
+    chunk_keys1 = set(
+        map(
+            lambda c: c.key,
+            (
+                subtasks[0].chunk_graph.results
+                + subtasks[1].chunk_graph.results
+                + subtasks[3].chunk_graph.results
+            ),
+        )
+    )
+    chunk_keys2 = set(
+        map(
+            lambda c: c.key,
+            (subtasks[2].chunk_graph.results + subtasks[4].chunk_graph.results),
+        )
+    )
+    assert chunk_keys1 == set(popped_seq[0:4])
+    assert chunk_keys2 == set(popped_seq[4:])
